Fixed reconnection logic, refactored connector and some other packages
