+++ packages/server/nodemon.json
@@ -1,8 +1,7 @@
 {
   "verbose": true,
-  "ignore": ["src/**/*.spec.ts"],
+  "ignore": ["*.spec.ts"],
   "watch": ["src/**/*.ts"],
-  "execMap": {
-    "ts": "node --inspect=0.0.0.0:9229 --nolazy -r ts-node/register"
-  }
+  "ext": "ts",
+  "exec": "ts-node --project tsconfig.json src/index.ts"
 }
