+++ packages/server/src/events/index.ts
@@ -1,0 +1,5 @@
+import EventEmitter from 'events'
+
+const eventEmitter = new EventEmitter()
+
+export default eventEmitter
