+++ packages/server/src/types/connector/index.d.ts
@@ -1,7 +1,7 @@
-import { DerivedStakingQuery } from '@polkadot/api-derive/types'
-import { HeaderExtended } from '@polkadot/api-derive/type'
+//import { DerivedStakingQuery } from
+// import { HeaderExtended } from '@polkadot/api-derive/type'
 
-export declare type SessionInfo = {
+declare type SessionInfo = {
   eraIndex: number
   eraLength?: number
   eraProgress?: number
@@ -12,48 +12,7 @@ export declare type SessionInfo = {
   sessionsPerEra?: number
 }
 
-export declare type EnhancedSlash = {
-  amount: string
-  who: string
-}
-
-export declare type BlockInfo = {
-  number: number
-  hash: string
-  timestamp: number
-}
-
-export interface EnhancedDerivedStakingQuery extends DerivedStakingQuery {
-  eraIndex: number
-  sessionIndex: number
-}
-
-export interface EventSlash {
-  accountId: string
-  amount: string
-}
-
-export interface EventReward {
-  amount: string
-}
-
-export interface EventOffence {
-  kind: string
-  timeSlot: string
-}
-
-export interface HeaderSessionInfo {
-  sessionIndex: number
-  eraIndex: number
-  slashes?: EventSlash[]
-  rewards?: EventReward[]
-  offences?: EventOffence[]
-}
-
-export interface EnhancedHeader {
-  author: string
-  number: number
-  hash: string
-  timestamp: number
-  sessionInfo: HeaderSessionInfo
-}
+// export declare type EnhancedSlash = {
+//   amount: string
+//   who: string
+// }
