+++ packages/server/tsconfig.json
@@ -10,16 +10,11 @@
     "experimentalDecorators": true,
     "target": "es6",
     "lib": ["es5", "es6", "dom"],
-    "module": "commonjs",
+    "module": "commonJS",
     "esModuleInterop": true,
     "sourceMap": true,
     "resolveJsonModule": true,
     "outDir": "./build",
-    "typeRoots": [
-      "./src/types",
-      "./node_modules/@polkadot/ts",
-      "./node_modules/@types",
-      "./node_modules/@polkadot/api-derive"
-    ]
+    "typeRoots": ["./node_modules/@types", "./src/types"]
   }
 }
